moved consts to const folder
